requirements.txt
# SysBot - AI Voice Assistant on Wheels
# Requirements for Raspberry Pi 5

# Core
numpy>=1.24.0

# Audio
pyaudio>=0.2.13
sounddevice>=0.4.6

# Wake Word (install separately on Pi with tflite-runtime)
# openwakeword>=0.5.0

# Speech-to-Text
faster-whisper>=0.9.0

# LLM
ollama>=0.1.0

# Text-to-Speech
piper-tts>=1.2.0
edge-tts>=6.1.0

# Display
pygame>=2.5.0

# Vision (Phase 2)
opencv-python>=4.8.0
mediapipe>=0.10.0

# GPIO (Phase 2 - Raspberry Pi only)
# gpiozero>=2.0.0
# pigpio>=1.78
# adafruit-circuitpython-pca9685>=3.4.0
# adafruit-circuitpython-servokit>=1.3.0

# API Server (Phase 3)
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Utilities
pyyaml>=6.0.0
